add a test case for a gvpr double free

See merge request graphviz/graphviz!4465

+++ cmd/smyrna/smyrnadefs.h
@@ -267,7 +267,7 @@ enum {SCHEMACOUNT = 4}; // number of colors
 	/*selected node color */
 	glCompColor selectedNodeColor;
 	/*default node alpha */
-	float defaultnodealpha;
+	double defaultnodealpha;
 
 	/*default line width */
 	float LineWidth;
